[UI - 8] : done edit action.
