+++ app/src/main/res/layout/task_dialog_fragment.xml
@@ -49,7 +49,7 @@
             android:layout_marginTop="@dimen/activity_vertical_margin"
             android:text="@string/priority_hint"
             android:textColor="@color/cardview_dark_background"
-            android:textSize="18dp" />
+            android:textSize="18sp" />
 
         <RadioGroup
             android:id="@+id/priority_group"
@@ -61,7 +61,7 @@
                 android:id="@+id/priority_low"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:checked="true"
+                android:checked="false"
                 android:text="@string/low_priority"
                 android:textColor="@color/cardview_dark_background" />
 
@@ -77,7 +77,7 @@
                 android:id="@+id/priority_high"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:checked="true"
+                android:checked="false"
                 android:text="@string/high_priority"
                 android:textColor="@color/cardview_dark_background" />
