+++ app/src/main/res/values/strings.xml
@@ -12,15 +12,16 @@
     <string name="high_priority">High</string>
     <string name="date_hint">Pick date</string>
     <string name="time_hint">Pick time</string>
-    <string name="exit_message">Discard new task?</string>
-    <string name="erase">erase</string>
+    <string name="exit_message">Discard changes?</string>
+    <string name="discard">discard</string>
     <string name="error_empty_title">type title</string>
     <string name="title_alert_toast">Please, enter task title!</string>
     <string name="date_time_alert_toast">Please, pick date and time!</string>
     <string name="task_created_toast">New task created.</string>
-    <string name="task_already_exist_toast">Task already exist!</string>
+    <string name="task_updated_toast">Task updated!</string>
     <string name="task_deleted_toast">Task deleted.</string>
     <string name="edit">Edit</string>
     <string name="delete">Delete</string>
+    <string name="edit_title">Edit task</string>
 
 </resources>
